refractor CustomDot to its own component
